refactor: cleaned up web/src structure
