+++ src/web/__tests__/assertBrowserEnvironment.ssr.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 // Because of the docblock above, jsdom test environment does not exist here
-import { assertBrowserEnvironment } from '../assertBrowserEnvironment';
+import { assertBrowserEnvironment } from '../utils/assertBrowserEnvironment';
 
 describe('assertBrowserEnvironment', () => {
   test('throws when DOM is missing', () => {
