+++ src/web/formats/EnrichedLink.ts
@@ -2,7 +2,7 @@ import Link, { type LinkOptions } from '@tiptap/extension-link';
 import { mergeAttributes, type CommandProps } from '@tiptap/core';
 import type { Editor } from '@tiptap/react';
 
-import { nativePosToTiptapPos } from '../positionMapping';
+import { nativePosToTiptapPos } from '../nativeMappers/positionMapping';
 import { isLinkBlocked } from './formatRules';
 import { findAutolinkRangesInWord } from '../pmPlugins/AutolinkPlugin/autolinkRegex';
