+++ src/web/tiptapWatchers/useOnChangeHtml.ts
@@ -1,8 +1,8 @@
 import { type Editor } from '@tiptap/react';
-import type { OnChangeHtmlEvent, SanitizationConfig } from '../types';
+import type { OnChangeHtmlEvent, SanitizationConfig } from '../../types';
 import type { NativeSyntheticEvent } from 'react-native';
 import { useOnEditorChange } from './useOnEditorChange';
-import { normalizeHtmlFromTiptap } from './normalization/tiptapHtmlNormalizer';
+import { normalizeHtmlFromTiptap } from '../normalization/tiptapHtmlNormalizer';
 
 export const useOnChangeHtml = (
   editor: Editor,
