+++ src/web/utils/assertBrowserEnvironment.ts
