+++ src/web/utils/pastedImageDimensions.ts
